Allows users to select desserts, amounts and payment options
This program allows a user to create a simulated dessert order through a command line. The user will be prompted to pick a dessert item, including the quantities. When the user is done with their order, they can create a new order, or exit the program and print a receipt. The program keeps track of orders, as well as the prices and tax for each order. In addition, the user can choose from different payments, including card, cash, or payment through the phone. This program demonstrates use cases for object oriented programming, including the use of abstract classes.

Hope you enjoy playing around with the program!